Move left non-deprecated types into Logbook.types.ts
