+++ frontend/src/features/Logbook/LegacyLogbook.types.ts
@@ -1,20 +1,4 @@
-import type { VesselIdentity } from '../../domain/entities/vessel/types'
 import type { Logbook } from '@features/Logbook/Logbook.types'
-
-export type VesselVoyage = {
-  endDate: string | null
-  isFirstVoyage: boolean
-  isLastVoyage: boolean
-  logbookMessagesAndAlerts: FishingActivities
-  startDate: string | null
-  tripNumber: string
-  vesselIdentity: VesselIdentity
-}
-
-export type FishingActivities = {
-  alerts: PNOAndLANWeightToleranceAlert[]
-  logbookMessages: LogbookMessage[]
-}
 
 /** @deprecated Replace with `Logbook.Message` . */
 // Can be done after using RTK in `getVesselLogbookFromAPI()` since undefined !== null.
@@ -49,26 +33,6 @@ export type LogbookMessage = {
   vesselName: string
 }
 
-export type PNOAndLANWeightToleranceAlert = {
-  externalReferenceNumber: string
-  id: string
-  internalReferenceNumber: string
-  ircs: string
-  value: PNOAndLANWeightToleranceAlertValue
-  vesselIdentifier: string
-  vesselName: string
-}
-
-export type PNOAndLANWeightToleranceAlertValue = {
-  catchesOverTolerance: PNOAndLANWeightToleranceAlertValueCatches[]
-  lanOperationNumber: string
-  minimumWeightThreshold: number
-  name: string
-  percentOfTolerance: number
-  pnoOperationNumber: string
-  type: 'PNO_LAN_WEIGHT_TOLERANCE_ALERT'
-}
-
 /** @deprecated Replace with `Logbook.Catch`. */
 // Can be done after using RTK in `getVesselLogbookFromAPI()` since undefined !== null.
 export type LogbookCatch = {
@@ -95,38 +59,4 @@ export type Gear = {
   gear: string | null
   gearName: string | null
   mesh: number | null
-}
-
-export type PNOAndLANWeightToleranceAlertValueCatches = {
-  lan: Object
-  pno: Object
-}
-
-export type ProtectedSpeciesCatch = {
-  careMinutes?: number
-  comment?: string
-  economicZone?: string
-  effortZone?: string
-  faoZone?: string
-  fate?: Fate
-  healthState?: HealthState
-  nbFish?: number
-  ring?: number
-  sex?: string
-  species: string
-  speciesName?: string
-  statisticalRectangle?: string
-  weight?: number
-}
-
-enum HealthState {
-  DEA,
-  ALI,
-  INJ
-}
-
-enum Fate {
-  DIS,
-  HEC,
-  DEA
 }
