+++ frontend/src/features/Logbook/Logbook.types.ts
@@ -1,6 +1,7 @@
-import type { ProtectedSpeciesCatch } from './LegacyLogbook.types'
+import type { LogbookMessage } from './LegacyLogbook.types'
 import type { AllSeafrontGroup, NoSeafrontGroup, SeafrontGroup } from '@constants/seafront'
 import type { PriorNotification } from '@features/PriorNotification/PriorNotification.types'
+import type { VesselIdentity } from 'domain/entities/vessel/types'
 
 export namespace Logbook {
   export type Message = PnoMessage | RetOperationMessage
@@ -202,6 +203,78 @@ export namespace Logbook {
     hasDesignatedPorts: boolean
     minimumNotificationPeriod: number
     pnoTypeName: string
+  }
+
+  export type PNOAndLANWeightToleranceAlert = {
+    externalReferenceNumber: string
+    id: string
+    internalReferenceNumber: string
+    ircs: string
+    value: PNOAndLANWeightToleranceAlertValue
+    vesselIdentifier: string
+    vesselName: string
+  }
+
+  export type PNOAndLANWeightToleranceAlertValue = {
+    catchesOverTolerance: PNOAndLANWeightToleranceAlertValueCatches[]
+    lanOperationNumber: string
+    minimumWeightThreshold: number
+    name: string
+    percentOfTolerance: number
+    pnoOperationNumber: string
+    type: 'PNO_LAN_WEIGHT_TOLERANCE_ALERT'
+  }
+
+  export type PNOAndLANWeightToleranceAlertValueCatches = {
+    lan: Object
+    pno: Object
+  }
+
+  export type ProtectedSpeciesCatch = {
+    careMinutes?: number
+    comment?: string
+    economicZone?: string
+    effortZone?: string
+    faoZone?: string
+    fate?: Fate
+    healthState?: HealthState
+    nbFish?: number
+    ring?: number
+    sex?: string
+    species: string
+    speciesName?: string
+    statisticalRectangle?: string
+    weight?: number
+  }
+
+  export type VesselVoyage = {
+    endDate: string | null
+    isFirstVoyage: boolean
+    isLastVoyage: boolean
+    logbookMessagesAndAlerts: FishingActivities
+    startDate: string | null
+    tripNumber: string
+    vesselIdentity: VesselIdentity
+  }
+
+  export type FishingActivities = {
+    alerts: PNOAndLANWeightToleranceAlert[]
+    logbookMessages: LogbookMessage[]
+  }
+
+  // ---------------------------------------------------------------------------
+  // Constants
+
+  enum HealthState {
+    DEA,
+    ALI,
+    INJ
+  }
+
+  enum Fate {
+    DIS,
+    HEC,
+    DEA
   }
 
   /* eslint-disable typescript-sort-keys/string-enum */
